refactor(api): split instruments out of hcapi
The many functions that just managed instrument data, without talking to
the backend or the robot config, now are in their own provider. This
leaves (primarily) cache_instruments in the hardware controller, since
it is a) specific to the ot2 or will be eventually, and b) needs to talk
to the backend, and an overload for plunger_position for ot3.
